Move TagMeta collection into TagModel metaclass
TagMeta was only applied to TagModel.tag_options when a field added it.
This led to ambiguity when multiple fields referred to the same model;
only the first's settings would be applied. This change moves the logic
into a new TagModelBase class (a metaclass for BaseTagModel) and raises
an exception when a field tries to set options for an existing tag
model.

However, south migrations can only store tag options on the fields, so
they're now given _set_tag_meta=True to tell the fields to force it
through.

Other changes:
* Added BaseTagTreeModel so south can correctly migrate tree models
* Added BaseTagModel._save_direct to let add_unique_column re-save
  models during a schemamigration without filling out auto-fields.
* Improved code consistency, removed redundant conditional branches