Tagged inline formsets work correctly
Fixes edge case when using a tagged model in an inline formset with a
tag model as the parent model.

Because BaseInlineFormSet expects the relationship to be a ForeignKey,
it sets the pk directly on field_id when adding fields (breaking tag
count), and expects the value_to_object of the ForeignKey form field
to be the pk (breaking edit).

This commit adds forms.TaggedInlineFormSet to fix both of these
issues, and automatically applies it in the admin site when needed.