Extend select2 adaptor to fix quoted tag selection
The tokenizer will generate correctly-quoted tags if the user enters commas or
spaces after the quotes, but the quotes will be included in the tag if someone
selects it from the dropdown (with click or enter).

We can't remove quotes in the tokenizer, as although that'll change the dropdown
value, it'll also change the user input, which we don't want to do.

This data adaptor extension will therefore trim unescaped quotes from the start
and end of a new tag in the dropdown so it's safe to click, without affecting the
input field.