Bugfix #164 Broken admin with Django 4.0
In Django 4.0 the `AutocompleteMixin` has been changed and the old
`AutocompleteMixin.media.fget(None)` will raise an (silent) error.

Fix this and add a test to check the included JS files.

Normally, the whole thing should be tested via browser.
But hopefully this way an error will be noticed faster.